+++ playwright.config.ts
@@ -6,6 +6,7 @@
 import { defineConfig } from '@playwright/test';
 import { CustomTestOptions } from './test/e2e/tests/_test.setup';
 import { currentsReporter, CurrentsFixtures, CurrentsWorkerFixtures } from '@currents/playwright';
+import * as fs from 'fs';
 
 // Merge Currents Fixtures into CustomTestOptions
 type ExtendedTestOptions = CustomTestOptions & CurrentsFixtures & CurrentsWorkerFixtures;
@@ -193,3 +194,18 @@ export default defineConfig<ExtendedTestOptions>({
 		},
 	],
 });
+
+/**
+ * Check if the current platform is openSUSE
+ */
+export function isOpenSUSE(): boolean {
+	try {
+		const osRelease = fs.readFileSync('/etc/os-release', 'utf8').toLowerCase();
+		const id = osRelease.match(/^id=(.*)$/m)?.[1]?.trim().replace(/^"|"$/g, '') ?? '';
+		const idLike = osRelease.match(/^id_like=(.*)$/m)?.[1]?.trim().replace(/^"|"$/g, '') ?? '';
+
+		return id.startsWith('opensuse') || id.includes('opensuse-leap') || idLike.includes('opensuse');
+	} catch {
+		return false;
+	}
+}
